testing on existence of rc-file in $HOME, before trying to copy it

+++ PKGBUILD
@@ -2,7 +2,7 @@
 
 pkgname=any-dl
 pkgver=0.9.6.b
-pkgrel=2
+pkgrel=3
 pkgdesc="Generic video downloader for principially any site."
 arch=('i686' 'x86_64')
 license=('GPL3')
@@ -25,6 +25,6 @@ cd ${srcdir}/${pkgname}-${pkgver}
 
 install -Dm 755 any-dl ${pkgdir}/usr/bin/any-dl         # install to Arch-Linux path
 
-cp -avf $HOME/.any-dl.rc $HOME/.any-dl.rc.bak           # backup of your any-dl-rc-file
+[ -f $HOME/.any-dl.rc  ] && cp -avf $HOME/.any-dl.rc $HOME/.any-dl.rc.bak           # backup of your any-dl-rc-file
 install -Dm 644 rc-file.adl ${pkgdir}$HOME/.any-dl.rc   # new any-dl-rc-file
 }
